+++ src/test/java/org/folio/rest/api/TestBaseWithInventoryUtil.java
@@ -72,6 +72,8 @@ public abstract class TestBaseWithInventoryUtil extends TestBase {
   protected static final UUID UUID_TEXT = UUID.fromString("6312d172-f0cf-40f6-b27d-9fa8feaf332f");
   protected static final UUID UUID_INSTANCE_TYPE = UUID.fromString("535e3160-763a-42f9-b0c0-d8ed7df6e2a2");
   protected static final UUID UUID_INSTANCE_DATE_TYPE = UUID.fromString("42dac21e-3c81-4cb1-9f16-9e50c81bacc4");
+  protected static final UUID UUID_INSTANCE_SUBJECT_TYPE_ID = UUID.fromString("d6488f88-1e74-40ce-81b5-b19a928ff5b1");
+  protected static final UUID UUID_INSTANCE_SUBJECT_SOURCE_ID = UUID.fromString("e894d0dc-621d-4b1d-98f6-6f7120eb0d40");
   protected static UUID journalMaterialTypeId;
   protected static String journalMaterialTypeID;
   protected static UUID bookMaterialTypeId;